+++ src/settings/Validation/engine/validators.js
@@ -18,6 +18,7 @@ import {
   hasLostItemProcessingFeeValue,
   hasPositiveReplacementAllowed,
   hasPatronBilledAfterAgedToLostValue,
+  hasPatronBilledAfterRecalledAgedToLostValue,
   hasChargeAmountItemSystemSelected,
   isToBeforeFrom,
   isDueAfterTo,
@@ -102,6 +103,10 @@ export default {
   hasPatronBilledAfterAgedToLostValue : {
     validate: hasPatronBilledAfterAgedToLostValue,
     message: <FormattedMessage id="ui-circulation.settings.lostItemFee.validate.hasPatronBilledAfterAgedToLostValue" />
+  },
+  hasPatronBilledAfterRecalledAgedToLostValue : {
+    validate: hasPatronBilledAfterRecalledAgedToLostValue,
+    message: <FormattedMessage id="ui-circulation.settings.lostItemFee.validate.hasPatronBilledAfterRecalledAgedToLostValue" />
   },
   hasPositiveLostItemProcessingFeeValue: {
     validate: isFloatGreaterThanZero,